netpulse/requirements.txt
Flask==3.1.0
dnspython==2.7.0
